requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Redis
redis==5.0.1

# LLM - LiteLLM 支持多种提供商
litellm>=1.48.0

# LLM Providers
openai==1.12.0  # OpenAI 和兼容接口
dashscope==1.14.1  # 阿里云千问 SDK（可选，litellm 已支持）

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.4.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
aiofiles==23.2.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx==0.26.0

# Development
black==24.1.1
ruff==0.2.0